fix(markers): run Intro & Credits like previews — respect GPU/CPU workers, same queue, retries and recovery
Intro & Credits now follows the same rules as previews at every step:
- GPU workers keep their work on the GPU. Text detection retries the GPU with a
  back-off (5 s doubling to 10 min) instead of switching to the CPU because it
  is faster. The per-GPU decode check is diagnostic only.
- Per-worker ffmpeg_threads apply to marker decodes, as they do to previews;
  CPU text helpers scale with cpu_threads. Nothing is hard-coded.
- Pause freezes marker FFmpeg runs, and a paused state survives settings saves
  (settings schema v19).
- Busy files wait with a bound and hand back to the retry queue; scheduled and
  Season jobs follow the server pin; restarts revive waiting jobs by downtime.
- A cancel never stores "nothing found".

Previews:
- BIFs are written atomically.
- Partial runs are sorted by cause. Interrupted runs (SIGTERM/255, SIGKILL, the
  stall watchdog, I/O errors) are discarded, and the Dolby Vision fallback tiers
  are skipped. Crashes and self-exits keep the 95% publish rule and the DV-safe
  last resort.
- Exit 251 with GPU error output is treated as a GPU failure, not a disk error.
- Fallback messages name the real cause.

Also fixes three integration and GPU tests that already failed on the base
commit: the gate journey teardown, the fixture timeouts, and the vendor skip
guard.

+++ media_preview_generator/processing/_shared.py
@@ -163,7 +163,7 @@ class _MediaServerProcessor:
         self,
         server_config: ServerConfig,
         *,
-        lookback_hours: int,
+        lookback_hours: float,
         library_ids: list[str] | None = None,
     ) -> Iterator[ProcessableItem]:
         """Vendor-specific recently-added query — subclasses must override.
